Keyboard: Move interfaces to interfaces/ subfolder
