+++ xbmc/games/addons/input/GameClientKeyboard.h
@@ -19,7 +19,7 @@
  */
 #pragma once
 
-#include "input/keyboard/IKeyboardHandler.h"
+#include "input/keyboard/interfaces/IKeyboardHandler.h"
 
 struct KodiToAddonFuncTable_Game;
