+++ xbmc/input/InputManager.h
@@ -33,7 +33,7 @@
 
 #include "Action.h"
 #include "windowing/XBMC_events.h"
-#include "input/keyboard/IKeyboardInputProvider.h"
+#include "input/keyboard/interfaces/IKeyboardInputProvider.h"
 #include "input/mouse/IMouseInputProvider.h"
 #include "input/KeyboardStat.h"
 #include "input/MouseStat.h"
